First quasi-working version of half type support. Some tests are still failing.

+++ GlmSharp/GlmSharp/Vec3/decvec3.glm.cs
@@ -416,7 +416,7 @@ namespace GlmSharp
         /// <summary>
         /// Returns a decvec3 from component-wise application of Degrees (Radians-To-Degrees Conversion).
         /// </summary>
-        public static decimal Degrees(decimal v) => v * 57.295779513082320876798154814105170332405472466564321m;
+        public static decimal Degrees(decimal v) => (decimal)(v * 57.295779513082320876798154814105170332405472466564321m);
         
         /// <summary>
         /// Returns a decvec3 from component-wise application of Radians (Degrees-To-Radians Conversion).
@@ -426,7 +426,7 @@ namespace GlmSharp
         /// <summary>
         /// Returns a decvec3 from component-wise application of Radians (Degrees-To-Radians Conversion).
         /// </summary>
-        public static decimal Radians(decimal v) => v * 0.0174532925199432957692369076848861271344287188854172m;
+        public static decimal Radians(decimal v) => (decimal)(v * 0.0174532925199432957692369076848861271344287188854172m);
         
         /// <summary>
         /// Returns a decvec3 from component-wise application of Acos ((decimal)Math.Acos((double)v)).
